2020-08-08 AFTER relocating the boolean
previously i had to create two versions of each item.  basically one to tell the planning system that it is WANTED, the other to tell the planning system that it is UNWANTED.  that was a lot of redundant work and code.  so i changed it so that this boolean is held outside of those old items, and is set when creating actions [actionItems, to be precise], not when creating items