2020-08-08 auto creates ''buy'' actions when needed
previously, i had to hand-craft a new action for ever kind of item that could be bought (and add this action to each NPCs list of "knownActions").  now, i only need to create the items.  now, when the planner sees that an item is needed [and has the stateCategory = "Inventory"], it will automatically create an action for buying that item, as one possible way to get it.