2021-08-01 NPC gives orders, even to player
previously, only the player could give orders to other "people" in the game.  now, the test NPC can do the same thing, including giving orders to the player.  this update [and the previous one] were part of my efforts to make the player and NPCs somewhat equivalent in terms of what they can do