2022-03-09 [F] factionState, planningState, ''soldier built''
introduced some more "state" variables.  this is because all planning could previously be achieved by looking into the inventory of the NPC doing the planning.  when NPCs need to coordinate in a faction, or make plans that involve items that are in storage rather than their own pockets, the items need to be tracked another way.  this system of multiple "states" is tricky because all of them must be updated, rather than just one, and sometimes there is some part of the code where i forget to update the others, which leads to unexpected behavior.