2022-03-24 [F] before merging tag scripts
i had two different scripts that both dealt with my tagging system.  one was for the system itself, the other was a quick way for me to add tags from inside the Unity editor.  these didn't really need to be separate, so i decided to merge them [but made this backup first]