plan patience, pending actions
some actions require the NPC to wait for some real-time events to unfold.  when this is the case, they should not discard their plan as if it is impossible.  this update adds a "pendingActions" list.  actions on the pendingActions list will still be considered part of the plan when determining whether or not the plan is defective.