faction growth goal action
created a simple way for a faction leader to have something like a "growth" type goal.  most of the desired state is specified in an action, because actions can store multiple prereqs, like multiple "goals".  the enaction phase modifies the prereqs/goals so that they are higher than they were previously.