AI won't try to shoot through walls
first checks if a raycast WOULD hit target, or if it would hit something else.

oddly, it was initially hitting their invisible "senseZone".  I thought i had fixed all of that ages ago, and i have no idea why the player was able to raycast through their senseZone, but they were not able to [except seemingly sometimes for targets that were not the player].  mystery.  but fixed now (had to set option in senseZone prefab to "IgnoreRaycast", which is why prefabs are included here)