ad-hoc plan to walk to gun, pick up, and fire it
NPCs [and player] no longer start with gun in their inventory.  a hand-coded/hard-wired plan is given to the NPCs to walk to a gun on the ground, "click" on it to pick it up, equip it, then repeatedly aim and fire it at random targets.

conditions are used to determine when to stop walking [proximity condition], and to reduce the firing rate of the shooting [with a "cooldown"].

the actual aiming of the gun seems to be malfunctioning.  they aim too high, and then too low, rather than directly at their chosen target (target selection seems to work correctly, they only target the player and other NPC during this step of the plan).  incorrect aiming behavior was also happening when i was working on tanks, but i fixed it.  not sure why this is happening (again?) now.