reworking gun cooldown to use numericalEffects etc
this seems very messy and overly complicated (though it mostly utilizes code i already had).  i might simplify this is the future, but it is good to see all those systems do work

firing a gun enacts TWO enactions:  the firing, and the reset of the cooling variable
cooldown is a "numerical variable" (accessed using an enum key in a dictionary) stored in the stateHolder [now called "interactable"]
firing is now conditioned on this dictionary value
every frame, interactables now implement any conditional effects if their conditions are met.  the condition for incrementing the cooldown is "autoCondition", which is always met