parallel enaction, error detection, NPC grabs gun
bunch of work for better parallel enaction system.  was previously limited to only allowing the NPC to aim and shoot ONCE [or fixed specified number of times] during the same timespan as their "walk to navpoint" enaction.  but obviously walking and shooting have to be way more independent of each other than that.

but NPCs kept failing to grab their shotguns.  so, did a LOT of very ad-hoc work to try to figure that out and fix it.  [eventually, found problems with raycast origin position, and adjusted the hitbox position for the shotgun in the prefab]

the debug tools are a large duct-taped mess, needs refactoring.  but that kind of problem-detection code will help with giving NPCs the ability to "rationally" CREATE plans [rather than merely FOLLOWING a plan that i write for them by hand].