beginning "animal game"
alright alright, I wasn't sure whether to store this on my main branch or whatever. but I guess I hope it'll all build towards the same larger game, and I want it to be coherent, so I might as well back it up all together. 

the main idea here that I'm working on is just getting some wildlife and such to work. and it seems again like it's leading me towards some good next steps for my code overall. 

I have a lot of backups on my computer, so I will be uploading them all in sequence. 

This backup was called “first bit of success”, it's from November 5th.