bit of object permanence sorta [nov 13th, 7:55pm]
animals would get stuck in a loop going back and forth it sees food from one position, but while going towards it it the food became hidden behind a wall or something. And so the NPC was hitting the “no visible food” condition to switch back to wandering [and they would have their previous Target, leading them to where they were standing when they first saw the food, so they just keep going back and forth forever]. 

So I added an ability for conditions to be “sticky”. that is, if they are true for one time they are checked, then they will stay “true” for a certain amount of time afterwards, even if it actually becomes false. 

It lasts for just a short amount of time allowing the animal to continue navigating towards the food target even when it is briefly hidden.