improved NPC combat behavior
Previously, the NPCs weren't very good at hitting their targets, and would oddly orbit each other very close. in order to get them to hit their targets properly, I had to offset their aim to account for the fact that they're gun is offset to the side of their center position. on its own that didn't work, so I had to separate their upper and lower bodies so that their aim was not being reset by the navmesh agent. the separation of the upper and lower bodies meant that their hitbox was only in their lower body. this made it difficult for them to hit each other. so I had to set the size and center of the hitbox so that it would encompass their entire body. 

I also changed the directions they would pick to move during combat, usually using the “apple pattern” which is a fairly effective dodge, but I added some random variety as well.