First RTS type command from NPC leader to NPC unit
the next step in making my survival mode game was for each team's units to start at their own base, and then for them to move across the map to attack enemy bases. I wasn't sure how to make that happen with code, until I realized it sounded a lot like an RTS Style command.  so I decided I might as well just add in RTS command enaction. 

So that means the player will be able to give commands to, at some point. I don't have the player's buttons or default command set yet.  or ways to select other types of commands.  but for now it will work for the survival mode game, and it will have room to grow.