first working illumination evaluation for stealth
this one feels good to have working.  uses "line of sight" for both guard's vision, and then also to light sources, which have a component that calculates illumination, which seems to accurately track in-game visibility [i might want to give the player a bit more margin for error, they currently need to be in near TOTAL darkness, but we'll see].

lots of new code classes.  but new means less messy/broken, so it wasn't too bad.