made an NPC go to random nearby hiding place
using light and line of sight to threats.  

but the raycast can hit the NPC's own body, so it can sometimes pick a point that is in its own "shadow", as if it's trying to hide behind itself for cover.  also, it doesn't take into account how visible its silhouette is